# api/openapi.yaml
openapi: 3.1.0
info:
  title: Reticora CMDB API
  version: 1.0.0
  description: |
    Multi-tenant Configuration Management Database API for MSPs and ISPs.
    Endpoints require tenant-scoped authentication unless explicitly marked otherwise.
  contact:
    name: DataHub Chiemgau
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/assignments:
    get:
      operationId: listAssignments
      tags: [Assignments]
      summary: List assignments
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: status
          in: query
          schema:
            type: string
        - name: assigned_to
          in: query
          schema:
            type: string
        - name: asset_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of assignments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createAssignment
      tags: [Assignments]
      summary: Create an assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentRequest'
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/assignments/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getAssignment
      tags: [Assignments]
      summary: Get an assignment
      responses:
        '200':
          description: Assignment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteAssignment
      tags: [Assignments]
      summary: Delete an assignment
      responses:
        '204':
          description: Assignment deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/assignments/{id}/return:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: returnAssignment
      tags: [Assignments]
      summary: Return an assigned asset or CI
      description: >-
        Marks the assignment as returned, recording the return condition and
        timestamp. The optional `notes` field overwrites the assignment notes
        when non-empty.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnAssignmentRequest'
      responses:
        '200':
          description: Assignment returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/assignments/{id}/transfer:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: transferAssignment
      tags: [Assignments]
      summary: Transfer an assignment to a new assignee
      description: >-
        Marks the existing assignment as transferred and creates a new active
        assignment of the same asset/CI for the new assignee.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferAssignmentRequest'
      responses:
        '201':
          description: New assignment created for the new assignee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/search:
    get:
      operationId: search
      tags: [Search]
      summary: Search tenant data
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 512 }
        - name: type
          in: query
          schema: { type: string }
          description: Comma-separated entity type filter.
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Maximum number of items to return.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: |
        Number of items to skip before returning data. Ignored when `cursor` is
        supplied.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: |
        Opaque keyset cursor taken from `next_cursor` of the previous page.
        Cursors are bound to the sort order they were issued for; sending one
        with a different `sort_by`/`sort_dir` combination yields 400.
    Search:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search term.
    SortBy:
      name: sort_by
      in: query
      schema:
        type: string
      description: Sort field name.
    SortDir:
      name: sort_dir
      in: query
      schema:
        type: string
        enum: [asc, desc]
      description: Sort direction.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    ProblemDetail:
      type: object
      description: RFC 7807 Problem Details object.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        trace_id:
          type: string
    PaginationEnvelope:
      type: object
      required: [total, limit, offset, has_more]
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
        next_cursor:
          type: string
          description: |
            Opaque cursor for the next page. Absent when no further rows follow.
...
    Assignment:
      type: object
      required: [id, organization_id, assigned_to, assigned_by, assignment_type, status, assigned_at, created_at, updated_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        asset_id:
          type: string
        ci_id:
          type: string
        assigned_to:
          type: string
        assigned_by:
          type: string
        assignment_type:
          type: string
        status:
          type: string
        assigned_at:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
        returned_at:
          type: string
          format: date-time
        return_condition:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAssignmentRequest:
      type: object
      required: [assigned_to]
      properties:
        asset_id:
          type: string
        ci_id:
          type: string
        assigned_to:
          type: string
        assignment_type:
          type: string
        due_date:
          type: string
          format: date-time
        notes:
          type: string
    AssignmentListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Assignment'
...
    ReturnAssignmentRequest:
      type: object
      description: Payload for returning an assigned asset or CI.
      properties:
        return_condition:
          type: string
        notes:
          type: string
    TransferAssignmentRequest:
      type: object
      required: [new_assignee]
      properties:
        new_assignee:
          type: string
        notes:
          type: string
...
    SearchHit:
      type: object
      required: [id, organization_id, entity_type, entity_id, title, url, score, updated_at]
      properties:
        id: { type: string }
        organization_id: { type: string }
        entity_type: { type: string, enum: [ci, asset, document, ticket, contact, compliance] }
        entity_id: { type: string }
        title: { type: string }
        summary: { type: string }
        url: { type: string }
        score: { type: number }
        highlights:
          type: array
          items: { type: string }
        metadata:
          type: object
          additionalProperties: { type: string }
        updated_at: { type: string, format: date-time }
    SearchResponse:
      type: object
      required: [data, total, limit, offset, has_more]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/SearchHit' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        has_more: { type: boolean }
...